+++ tests/src/shared/inverter/tb.rs
@@ -6,7 +6,8 @@ use serde::{Deserialize, Serialize};
 use sky130pdk::corner::Sky130Corner;
 use sky130pdk::Sky130CommercialPdk;
 use spectre::blocks::{Pulse, Vsource};
-use spectre::{Options, Spectre, Tran};
+use spectre::tran::Tran;
+use spectre::{Options, Spectre};
 use substrate::block::Block;
 use substrate::context::Context;
 use substrate::io::TestbenchIo;
