+++ tests/src/shared/vdivider/tb.rs
@@ -2,7 +2,8 @@ use rust_decimal::prelude::ToPrimitive;
 use rust_decimal_macros::dec;
 use serde::{Deserialize, Serialize};
 use spectre::blocks::{Iprobe, Vsource};
-use spectre::{Options, Spectre, Tran, TranCurrent, TranCurrentKey, TranVoltage, TranVoltageKey};
+use spectre::tran::{Tran, TranCurrent, TranCurrentKey, TranVoltage, TranVoltageKey};
+use spectre::{Options, Spectre};
 use substrate::block::Block;
 use substrate::io::Signal;
 use substrate::io::TestbenchIo;
