+++ tests/src/sim/spectre.rs
@@ -8,7 +8,8 @@ use serde::{Deserialize, Serialize};
 use sky130pdk::corner::Sky130Corner;
 use sky130pdk::Sky130CommercialPdk;
 use spectre::blocks::Vsource;
-use spectre::{Options, Spectre, Tran};
+use spectre::tran::Tran;
+use spectre::{Options, Spectre};
 use substrate::block::Block;
 use substrate::cache::Cache;
 use substrate::context::Context;
